requirements.txt
discord.py
facebook-scraper